Removed `Rectangled` to its own module
